Move texture/texture.rs to texture/mod.rs
